debug win32 version(not complete)/update TODO
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2591 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/include/connection.h
@@ -172,7 +172,7 @@ protected:
 	int forceParsing;
 
 	/*! Connection priority, used by the scheduler.  */
-	u_long priority;
+	int priority;
 
 	/*! Libevent event, used by the connections scheduler.  */
 	event ev;
